Rename msTeams directory to ms-teams
