+++ packages/client/src/clients/ms-teams/interfaces.ts
